+++ backend/src/routes/userRoutes.ts
@@ -1,13 +1,12 @@
 import { Router } from 'express';
-import { createUser, loginUser, getCurrentUser } from '../controllers/userController';
+import { createUser, getCurrentUser } from '../controllers/userController';
 import validate from '../middleware/validate';
-import { createUserSchema, loginUserSchema } from '../schemas/userSchema';
+import { createUserSchema } from '../schemas/userSchema';
 import auth from '../middleware/auth';
 
 const router = Router();
 
-router.post('/register', validate(createUserSchema), createUser);
-router.post('/login', validate(loginUserSchema), loginUser);
+router.post('/', validate(createUserSchema), createUser);
 router.get('/me', auth, getCurrentUser);
 
 export default router;