+++ frontend/src/main.tsx
@@ -2,9 +2,12 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
 import './index.css';
+import { ThemeProvider } from './context/ThemeContext'; // Import ThemeProvider
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <App />
+    <ThemeProvider> {/* Wrap App with ThemeProvider */}
+      <App />
+    </ThemeProvider>
   </React.StrictMode>
 );